# docs/openapi-pos-mvp.yaml
openapi: 3.0.3
...
info:
  title: BoothPOS API
  version: 1.4.0
  description: |
    API internal untuk BoothPOS — sistem POS berbasis event untuk toko
    merchandise yang berjualan online maupun di bazaar/event. Berjalan di
    localhost pada perangkat operasional toko, tanpa akses publik,
    sejalan dengan model lisensi sekali bayar dan instalasi lokal per toko.

    **Autentikasi** — token bearer. Token diperoleh dari `/auth/login`
    dan dikirim pada header `Authorization: Bearer <token>`.

    **Format uang** — seluruh nilai uang berupa string desimal dua angka
    di belakang koma, contoh `"150000.00"`. String dipakai agar tidak ada
    galat pembulatan floating point di sisi klien.

    **Zona waktu** — seluruh timestamp dalam format ISO 8601 dengan
    offset waktu lokal perangkat.
servers:
  - url: http://localhost:8000/api/v1
    description: Instalasi lokal pada laptop operasional
...
security:
  - bearerAuth: []

# =====================================================================
# PATHS
# =====================================================================
paths:
...
  /settings:
    get:
      tags: [Settings]
      summary: Daftar seluruh pengaturan
      description: |
        Hanya owner/admin (F14 — Setting & config management). Tabel
        `payment_channels` (nomor rekening) TIDAK ikut di sini; itu
        resource terpisah dengan penyamaran nomor rekeningnya sendiri.
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Setting' }
        '403': { $ref: '#/components/responses/Forbidden' }
    put:
      tags: [Settings]
      summary: Ubah satu atau beberapa pengaturan sekaligus
      description: |
        Bentuk BULK, bukan satu key per request — layar pengaturan
        biasanya menyimpan beberapa field sekaligus (nama toko, kontak,
        format struk). Ini juga satu-satunya jalur admin untuk mengubah
        `multi_artist_enabled` (upgrade lisensi Pro ke Master); lihat
        catatan lisensi pada `POST /artists`. `type`/`group` boleh
        dikosongkan pada key yang sudah ada — nilai lama dipakai lagi.
        Setiap key yang berubah tercatat di log aktivitas (F13.4).

        Juga satu-satunya jalur mengubah `system_mode`
        (003-seed-demo-live) — key baru bernilai `demo`/`live`, group
        `system`. Tidak ada endpoint terpisah untuk ini (lihat
        specs/003-seed-demo-live/contracts/settings-system-mode.md):

        ```json
        { "settings": [{ "key": "system_mode", "value": "demo", "type": "string", "group": "system" }] }
        ```

        006-purchase-order-and-ops (US7) — juga satu-satunya jalur untuk
        tiga key baru ini, tidak ada endpoint terpisah: `theme_accent_color`
        (hex warna, group `appearance`, diterapkan klien sebagai override
        CSS custom property `--color-brand`), `receipt_footer_text` (group
        `receipt`, dirender di struk hanya bila terisi — dihilangkan sama
        sekali, bukan blok kosong, saat kosong), `receipt_show_logo`
        (boolean-as-string, group `receipt`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [settings]
              properties:
                settings:
                  type: array
                  minItems: 1
                  items: { $ref: '#/components/schemas/SettingInput' }
      responses:
        '200':
          description: Diperbarui
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Setting' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
...
    Forbidden:
      description: Peran tidak memiliki izin
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationError:
      description: Validasi gagal
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
              errors:
                type: object
                additionalProperties:
                  type: array
                  items: { type: string }
  schemas:
    Error:
      type: object
      properties:
        message: { type: string }
        code: { type: string, nullable: true, example: INSUFFICIENT_STOCK }
...
    Setting:
      type: object
      properties:
        key: { type: string, example: store_name }
        value:
          type: string
          nullable: true
          description: Selalu disimpan sebagai string mentah, terlepas dari `type`. Interpretasinya (mis. boolean) terjadi di sisi pembaca.
        type: { type: string, enum: [string, integer, decimal, boolean, json] }
        group: { type: string, example: receipt }
    SettingInput:
      type: object
      required: [key, value]
      properties:
        key: { type: string, maxLength: 100, example: store_name }
        value:
          nullable: true
          description: "Boleh string, angka, boolean, atau objek (untuk type=json) — dinormalisasi ke string sebelum disimpan."
          example: "Toko Merchandise Dummy"
        type:
          type: string
          enum: [string, integer, decimal, boolean, json]
          description: Dikosongkan berarti memakai `type` dari baris yang sudah ada, atau `string` bila key baru.
        group:
          type: string
          maxLength: 50
          description: Dikosongkan berarti memakai `group` dari baris yang sudah ada, atau `general` bila key baru.
...